Split up project into main file and util files
